+++ main.py
@@ -1,0 +1,18 @@
+from __future__ import annotations
+
+import sys
+from PyQt6.QtWidgets import QApplication
+
+from gui import MainWindow
+
+
+def main() -> None:
+    app = QApplication(sys.argv)
+    window = MainWindow()
+    window.resize(1200, 800)
+    window.show()
+    sys.exit(app.exec())
+
+
+if __name__ == "__main__":
+    main()
